+++ src/App/Documentation/Components/Panel/index.js
@@ -10,19 +10,26 @@ const Overview = () => (
         <h2 id="overview">Overview</h2>
         <p>Panel...</p>
         <ComponentPreview language="html" showCasePanel={true} codeFigure={true} removeOuterTag={false}>
-            <PanelComponent type="brand" title="Panel title" subTitle="Panel sub title">
+            <PanelComponent
+                title="Panel title"
+                subTitle="Panel subtitle"
+                footerText="Footer content goes here."
+                footerBtnText="Footer button"
+                footerCentered={true}
+            >
                 <p>Content here.</p>
             </PanelComponent>
         </ComponentPreview>
     </div>
 );
 
-const Demo = () => (
+const PanelHeaders = () => (
     <div>
-        <h2 id="demo">Demo</h2>
-        <p>Default Panel</p>
+        <h2 id="panel-headers">Panel headers</h2>
+        <p>Two types of headers... <code className="token property">.panel-default</code> and <code className="token property">.panel-brand</code>...</p>
         <ComponentPreview language="html" showCasePanel={true} codeFigure={true} removeOuterTag={false}>
-            <PanelComponent />
+            <PanelComponent title="Default panel" />{"\n\n"}
+            <PanelComponent type="brand" title="Panel brand" />
         </ComponentPreview>
     </div>
 );
@@ -31,7 +38,7 @@ const PanelText = () => (
     <div className="col-md-12 col-lg-10 doc-body">
         <p className="lead">Panels...</p>
         <Overview />
-        {/* <Demo /> */}
+        <PanelHeaders />
     </div>
 );